Refactor sync logic into abstract class
