+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-12-23T06:44:38.574Z\n"
-"PO-Revision-Date: 2019-12-23T06:44:38.574Z\n"
+"POT-Creation-Date: 2019-12-23T07:33:16.467Z\n"
+"PO-Revision-Date: 2019-12-23T07:33:16.467Z\n"
 
 msgid "Search by name"
 msgstr ""
@@ -347,19 +347,10 @@ msgstr ""
 msgid "Deleting objects in instance {{instance}}"
 msgstr ""
 
-msgid "Fetching metadata from origin instance"
+msgid "Preparing synchronization"
 msgstr ""
 
 msgid "Importing data in instance {{instance}}"
-msgstr ""
-
-msgid "Fetching data from data sets and data element groups"
-msgstr ""
-
-msgid "Fetching data from data elements"
-msgstr ""
-
-msgid "Fetching data from events"
 msgstr ""
 
 msgid "Not a valid DHIS2 instance"
