+++ src/pages/sync-rules-list/SyncRulesListPage.jsx
@@ -15,11 +15,9 @@ import Dropdown from "../../components/dropdown/Dropdown";
 import PageHeader from "../../components/page-header/PageHeader";
 import SharingDialog from "../../components/sharing-dialog/SharingDialog";
 import SyncSummary from "../../components/sync-summary/SyncSummary";
-import {
-    startAggregatedSynchronization,
-    startEventsSynchronization,
-    startMetadataSynchronization,
-} from "../../logic/synchronization";
+import { AggregatedSync } from "../../logic/sync/aggregated";
+import { EventsSync } from "../../logic/sync/events";
+import { MetadataSync } from "../../logic/sync/metadata";
 import Instance from "../../models/instance";
 import SyncReport from "../../models/syncReport";
 import SyncRule from "../../models/syncRule";
@@ -35,15 +33,15 @@ import { getValidationMessages } from "../../utils/validations";
 const config = {
     metadata: {
         title: i18n.t("Metadata Synchronization Rules"),
-        action: startMetadataSynchronization,
+        SyncClass: MetadataSync,
     },
     aggregated: {
         title: i18n.t("Aggregated Synchronization Rules"),
-        action: startAggregatedSynchronization,
+        SyncClass: AggregatedSync,
     },
     events: {
         title: i18n.t("Events Synchronization Rules"),
-        action: startEventsSynchronization,
+        SyncClass: EventsSync,
     },
 };
 
@@ -202,16 +200,14 @@ class SyncRulesPage extends React.Component {
         history.push(`/sync-rules/${match.params.type}/edit/${rule.id}`);
     };
 
-    executeRule = async ({ builder, name, id, type = "metadata" }) => {
+    executeRule = async ({ builder, name, id: syncRule, type = "metadata" }) => {
         const { d2, loading } = this.props;
         const { api } = this.context;
-        const { action } = config[type];
+        const { SyncClass } = config[type];
 
         try {
-            for await (const { message, syncReport, done } of action(d2, api, {
-                ...builder,
-                syncRule: id,
-            })) {
+            const sync = new SyncClass(d2, api, { ...builder, syncRule });
+            for await (const { message, syncReport, done } of sync.execute()) {
                 if (message) loading.show(true, message);
                 if (syncReport) await syncReport.save(d2);
                 if (done && syncReport) {
